Moved damage parsing out of t_parser.cpp and in to t_damage.cpp as an overloaded constructor.
This way the source file defining the event type is also responsible for
parsing the message from the log.

The parser will now just detect which channel the message is coming
from, create the appropriate object and pass the message to it.

It may be required to do some parsing on message to detect the event
type when different events come in on the same channel.